fix: move react dependencies to be peer dependencies (#91)
This PR moves react and react-dom to be peer dependencies, solving issues like #90 where there is a mismatch in React versions between the library and the host.
This kind of move is usually a breaking change - but since we can assume that all the consumers of the React SDK already have React installed in their dependencies (since they need to render <UIResourceRenderer> somehow) - this might not be a problem.